# Remove duplicate NavigationTabBar from Home tab on mobile