--- a/src/pages/home/sidebar/BaseSidebarScreen.native.tsx
+++ b/src/pages/home/sidebar/BaseSidebarScreen.native.tsx
@@ -1,18 +1,12 @@
 import React, {useEffect} from 'react';
-import NavigationTabBar from '@components/Navigation/NavigationTabBar';
-import NAVIGATION_TABS from '@components/Navigation/NavigationTabBar/NAVIGATION_TABS';
 import ScreenWrapper from '@components/ScreenWrapper';
-import useResponsiveLayout from '@hooks/useResponsiveLayout';
 import useThemeStyles from '@hooks/useThemeStyles';
-import {isMobile} from '@libs/Browser';
 import Performance from '@libs/Performance';
 import CONST from '@src/CONST';
 import SidebarInboxContent from './SidebarInboxContent';
 
 function BaseSidebarScreen() {
     const styles = useThemeStyles();
-    const {shouldUseNarrowLayout} = useResponsiveLayout();
-    const shouldDisplayLHB = !shouldUseNarrowLayout;
 
     useEffect(() => {
         Performance.markStart(CONST.TIMING.SIDEBAR_LOADED);
@@ -21,13 +15,12 @@ function BaseSidebarScreen() {
     return (
         <ScreenWrapper
             shouldEnableKeyboardAvoidingView={false}
-            style={[styles.sidebar, isMobile() ? styles.userSelectNone : {}]}
+            style={[styles.sidebar, styles.userSelectNone]}
             testID={BaseSidebarScreen.displayName}
-            bottomContent={!shouldDisplayLHB && <NavigationTabBar selectedTab={NAVIGATION_TABS.HOME} />}
         >
             {({insets}) => (
                 <SidebarInboxContent
-                    shouldUseNarrowLayout={shouldUseNarrowLayout}
+                    shouldUseNarrowLayout
                     insets={insets}
                 />
             )}