--- a/src/styles/index.ts
+++ b/src/styles/index.ts
@@ -631,6 +631,10 @@ const staticStyles = (theme: ThemeColors) =>
             backgroundColor: theme.appBG,
         },
 
+        scrollPaddingWhenNoBottomTab: {
+            paddingBottom: variables.bottomTabHeight,
+        },
+
         navigationTabBarItem: {
             height: '100%',
             display: 'flex',